Light refactoring and reorganization
