--- packages/nx-plugin/src/generators/eslint/index.ts
@@ -1,1 +1,0 @@
-export { eslintGenerator } from "./generator";
