+++ packages/nx-plugin/src/generators/linting/eslint/eslintGenerator.ts
