+++ packages/nx-plugin/src/generators/linting/eslint/files/.eslintrc.json
