+++ packages/nx-plugin/src/generators/linting/eslint/index.ts
@@ -1,0 +1,2 @@
+export * from "./eslintGenerator";
+export { eslintGenerator as default } from "./eslintGenerator";
