+++ packages/nx-plugin/src/generators/linting/eslint/schema.json
@@ -1,0 +1,7 @@
+{
+  "$schema": "http://json-schema.org/schema",
+  "cli": "nx",
+  "$id": "EsLint",
+  "title": "",
+  "type": "object"
+}
