+++ packages/nx-plugin/src/generators/linting/index.ts
@@ -1,0 +1,2 @@
+export * from "./lintingGenerator";
+export { lintingGenerator as default } from "./lintingGenerator";
