+++ packages/nx-plugin/src/generators/linting/lintingGenerator.ts
@@ -1,0 +1,7 @@
+import { Tree } from "@nrwl/devkit";
+
+import { eslintGenerator } from "./eslint";
+
+export function lintingGenerator(tree: Tree) {
+  eslintGenerator(tree);
+}
