+++ packages/nx-plugin/src/generators/linting/schema.json
@@ -1,0 +1,7 @@
+{
+  "$schema": "http://json-schema.org/schema",
+  "cli": "nx",
+  "$id": "Linting",
+  "title": "",
+  "type": "object"
+}
