+++ packages/nx-plugin/src/generators/preset/index.ts
@@ -1,0 +1,3 @@
+export * from "./presetGenerator";
+export { presetGenerator as default } from "./presetGenerator";
+export * from "./presetGenerator.schema";
