+++ packages/nx-plugin/src/generators/preset/presetGenerator.schema.ts
